Cleanup alias generation into its own file
